--- docs/src/app/(public)/[...not-found]/page.tsx
@@ -1,7 +1,0 @@
-import { notFound } from 'next/navigation';
-
-// https://github.com/vercel/next.js/discussions/50034
-
-export default function NotFound() {
-  notFound();
-}
